- windows devel
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@499 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/av/FormatInputStream.cpp
@@ -51,7 +51,7 @@ namespace org {
                 return (AVInputStream*) formatCtx->streams[streamIndex];
             }
 
-            int FormatInputStream::available(bool withBlocking) {
+            long long int FormatInputStream::available(bool withBlocking) {
                 return 0;
             }
