+++ src/org/esb/io/FileInputStream.h
@@ -40,7 +40,7 @@ public:
                 FileInputStream(File * file);
                 FileInputStream(const std::string);
                 ~FileInputStream();
-                int available(bool i=false);
+                long long int available(bool i=false);
                 int read(string & str);
                 int read(unsigned char * buffer, int length);
                 int read(vector<unsigned char>&buffer);
