+++ src/org/esb/io/ObjectInputStream.cpp
@@ -2,7 +2,7 @@
 
 using namespace org::esb::io;
 		ObjectInputStream::ObjectInputStream(InputStream * is):_is(is){};
-		int ObjectInputStream::available(bool isBlocking){return 0;};
+		long long int ObjectInputStream::available(bool isBlocking){return 0;};
 		int ObjectInputStream::read(unsigned char * buffer, int length){return 0;};
 		int ObjectInputStream::read(vector<unsigned char>&buffer){return 0;};
 		int ObjectInputStream::read(){return 0;};
